QuadLinked node and TRIE implementation.
An initial version of the quad linked trie and node has been implemented
to satisfy the previously created interfaces. A simple benchmarking app
has been added to compare the insertion and searching times of the TRIE
and the unordered_map provided by the STL.

A resources directory has been added to the source structure so that a
words file for the benchmark has a proper place in the project.

A simple makefile for the project is included but it has to be improved.