Documentation of QuadLinkedTrieNode class.
A new directory doc for doxygen documentation has been included in the
project's source tree. The doxygen.cfg file is configured to generate
the documentation of the files in the directories /lib and /include.

The class header for the QuadLinkedTrieNode has been fully documented,
the interface is completed but the implementation is not commented yet.
The documentation for that class has been generated with doxygen tool.

What's more, the names of the parameters and class members have been
modified according to a common convention, adding a m character before
a class member name and a p one before a paramenter name.